requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
kombu==5.3.7
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
pyTelegramBotAPI==4.18.1
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.2
six==1.16.0
sqlparse==0.5.0
telebot==0.0.5
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
